+++ dataservices/management/commands/helpers.py
@@ -196,3 +196,141 @@ def ingest_data(engine, metadata, on_before_visible, batches):
             upsert=pg_bulk_ingest.Upsert.OFF,
             delete=pg_bulk_ingest.Delete.BEFORE_FIRST_BATCH,
         )
+
+
+def get_dbtsector_table_batch(data, data_table):
+    table_data = (
+        (
+            data_table,
+            (
+                dbt_sector['id'],
+                dbt_sector['field_01'],
+                dbt_sector['full_sector_name'],
+                dbt_sector['sector_cluster__april_2023'],
+                dbt_sector['field_04'],
+                dbt_sector['field_05'],
+                dbt_sector['field_02'],
+            ),
+        )
+        for dbt_sector in data
+    )
+
+    return (
+        None,
+        None,
+        table_data,
+    )
+
+
+def get_dbtsector_postgres_table(metadata):
+    return sa.Table(
+        "dataservices_dbtsector",
+        metadata,
+        sa.Column("id", sa.INTEGER, nullable=False),
+        sa.Column("sector_id", sa.TEXT, nullable=True),
+        sa.Column("full_sector_name", sa.TEXT, nullable=True),
+        sa.Column("sector_cluster_name", sa.TEXT, nullable=True),
+        sa.Column("sector_name", sa.TEXT, nullable=True),
+        sa.Column("sub_sector_name", sa.TEXT, nullable=True),
+        sa.Column("sub_sub_sector_name", sa.TEXT, nullable=True),
+        sa.Index(None, "id"),
+        schema="public",
+    )
+
+
+def get_sectors_gva_value_bands_table(metadata):
+    return sa.Table(
+        "dataservices_sectorgvavalueband",
+        metadata,
+        sa.Column("id", sa.INTEGER, nullable=False),
+        sa.Column("full_sector_name", sa.TEXT, nullable=False),
+        sa.Column("value_band_a_minimum", sa.INTEGER, nullable=False),
+        sa.Column("value_band_b_minimum", sa.INTEGER, nullable=False),
+        sa.Column("value_band_c_minimum", sa.INTEGER, nullable=False),
+        sa.Column("value_band_d_minimum", sa.INTEGER, nullable=False),
+        sa.Column("value_band_e_minimum", sa.INTEGER, nullable=False),
+        sa.Column("start_date", sa.DATE, nullable=False),
+        sa.Column("end_date", sa.DATE, nullable=False),
+        sa.Column("sector_classification_value_band", sa.TEXT, nullable=False),
+        sa.Column("sector_classification_gva_multiplier", sa.TEXT, nullable=False),
+        schema="public",
+    )
+
+
+def get_sectors_gva_value_bands_batch(data, data_table):
+    table_data = (
+        (
+            data_table,
+            (
+                sectors_gva_value_bands['id'],
+                sectors_gva_value_bands['full_sector_name'],
+                sectors_gva_value_bands['value_band_a_minimum'],
+                sectors_gva_value_bands['value_band_b_minimum'],
+                sectors_gva_value_bands['value_band_c_minimum'],
+                sectors_gva_value_bands['value_band_d_minimum'],
+                sectors_gva_value_bands['value_band_e_minimum'],
+                sectors_gva_value_bands['start_date'],
+                sectors_gva_value_bands['end_date'],
+                sectors_gva_value_bands['sector_classification_value_band'],
+                sectors_gva_value_bands['sector_classification_gva_multiplier'],
+            ),
+        )
+        for sectors_gva_value_bands in data
+    )
+
+    return (
+        None,
+        None,
+        table_data,
+    )
+
+
+def get_investment_opportunities_data_table(metadata):
+
+    return sa.Table(
+        "dataservices_dbtinvestmentopportunity",
+        metadata,
+        sa.Column("id", sa.INTEGER, nullable=False),
+        sa.Column("opportunity_title", sa.TEXT),
+        sa.Column("description", sa.TEXT),
+        sa.Column("nomination_round", sa.FLOAT),
+        sa.Column("launched", sa.BOOLEAN),
+        sa.Column("opportunity_type", sa.TEXT),
+        sa.Column("location", sa.TEXT),
+        sa.Column("sub_sector", sa.TEXT),
+        sa.Column("levelling_up", sa.BOOLEAN),
+        sa.Column("net_zero", sa.BOOLEAN),
+        sa.Column("science_technology_superpower", sa.BOOLEAN),
+        sa.Column("sector_cluster", sa.TEXT),
+        schema="public",
+    )
+
+
+def get_investment_opportunities_batch(data, data_table):
+
+    table_data = (
+        (
+            data_table,
+            (
+                investment_opportunity['id'],
+                investment_opportunity['opportunity_title'],
+                investment_opportunity['description'],
+                investment_opportunity['nomination_round'],
+                investment_opportunity['launched'],
+                investment_opportunity['opportunity_type'],
+                investment_opportunity['location'],
+                investment_opportunity['sub_sector'],
+                investment_opportunity['levelling_up'],
+                investment_opportunity['net_zero'],
+                investment_opportunity['science_technology_superpower'],
+                investment_opportunity['sector_cluster'],
+            ),
+        )
+        for investment_opportunity in data
+    )
+
+    return (
+        None,
+        None,
+        table_data,
+    )
