+++ dataservices/tests/test_ingestion_pattern.py
@@ -7,21 +7,10 @@ from django.core import management
 from django.test import override_settings
 from sqlalchemy.future.engine import Engine
 
-from dataservices.management.commands.import_dbt_investment_opportunities import (
-    get_investment_opportunities_batch,
-    get_investment_opportunities_data_table,
-    save_investment_opportunities_data,
-)
-from dataservices.management.commands.import_dbt_sectors import (
-    get_dbtsector_postgres_table,
-    get_dbtsector_table_batch,
-    save_dbt_sectors_data,
-)
-from dataservices.management.commands.import_sectors_gva_value_bands import (
-    get_sectors_gva_value_bands_batch,
-    get_sectors_gva_value_bands_table,
-    save_sectors_gva_value_bands_data,
-)
+from dataservices.management.commands.import_dbt_investment_opportunities import save_investment_opportunities_data
+from dataservices.management.commands.import_dbt_sectors import save_dbt_sectors_data
+from dataservices.management.commands.import_sectors_gva_value_bands import save_sectors_gva_value_bands_data
+from dataservices.management.commands import helpers
 
 dbsector_data = [
     {
@@ -157,7 +146,7 @@ def test_save_dbtsector_data(mock_connection, mock_ingest, dbtsector_data):
 @pytest.mark.django_db
 def test_get_dbtsector_table_batch(dbtsector_data):
     metadata = sa.MetaData()
-    ret = get_dbtsector_table_batch(dbtsector_data, get_dbtsector_postgres_table(metadata))
+    ret = helpers.get_dbtsector_table_batch(dbtsector_data, helpers.get_dbtsector_postgres_table(metadata))
     assert next(ret[2]) is not None
 
 
@@ -174,7 +163,9 @@ def test_save_sectors_gva_value_bands_data(mock_connection, mock_ingest, sectors
 @pytest.mark.django_db
 def test_get_sectors_gva_value_bands_batch(sectors_gva_value_bands_data):
     metadata = sa.MetaData()
-    ret = get_sectors_gva_value_bands_batch(sectors_gva_value_bands_data, get_sectors_gva_value_bands_table(metadata))
+    ret = helpers.get_sectors_gva_value_bands_batch(
+        sectors_gva_value_bands_data, helpers.get_sectors_gva_value_bands_table(metadata)
+    )
     assert next(ret[2]) is not None
 
 
@@ -191,7 +182,7 @@ def test_save_investment_opportunities_data(mock_connection, mock_ingest, invest
 @pytest.mark.django_db
 def test_get_investment_opportunities_batch(investment_opportunities_data):
     metadata = sa.MetaData()
-    ret = get_investment_opportunities_batch(
-        investment_opportunities_data, get_investment_opportunities_data_table(metadata)
+    ret = helpers.get_investment_opportunities_batch(
+        investment_opportunities_data, helpers.get_investment_opportunities_data_table(metadata)
     )
     assert next(ret[2]) is not None
